We have made lots of changes. There is now a button for the update theme notice which apparently returns done for when updating but does not work.
We have also fixed lots of small issues, all the hooks are working in the theme. We also have updated the README to explain this. Aisis is built on templates, Modules and Hooks with loaders and registiring doing the lifting.

The short codes popup on the post page for the edit or new post has some new tabs and info regarding images and general post content.

We added a new method to the file handeling class which returns all files in a directory that match the extension we pass in.

All editors now have appropriate messages, popups and what not for Empty, same as what was loaded and for new content being submited when a user hits save.