Aisis Admin panel now sees the inclusion of the Options package in the AisisCore->AdminPanel->Modules package. This options package is page specific and uses the get_settings_fields($page, $section); to render out a section.
By doing this I am able to have multiple textareas on the same page each with there own validation or each with the same validation that checks all of them. Currently there is no validation hooked up to the Aisis Core Options.

You will also notice that instead of aisis_options the page is now aisis-core-options.