Aisis Admin Pannel Options now saves to the options table and validates upon empty text areas. Our next step is to validate that no html has been input in fields that dont allow for html.
This is only working currently on the Aisis Core Options Page.

If the content is left empty then the appropriate error(s) are thrown and appropriate Js executes. If one is empty but all are filled in then what we have is an error, the others that are not empty are not saved and the empty field is repopulated with default content.

TODO: is there a better way to do this?