The error reporting and success messages are now working properly in Aisis for the Admin Core Options. I have started working on the Aisis css Editor, Essentially we are saving data to a file that should exist at said location. currently if said file does not exist then we do not save the file, we throw an error.
TODO: essential logic is that:

if the contents in the options table and the contents of the input are the exact ame do nothing but show a notice.
if the contents in the input are empty, throw a notice but save to the options table.

There should be checking for directory existence, file existance and so on.

We also added a dubugger which takes in the thig you want printd out and weather or not we should echo it in pretty format or not.