Lots of changes went into this release for the development branch. We even closed out some tickets and cleaned up some code, files and other things. One of the major things you will notice is the twitter widget has been redone, simplified and made better.
We have also cleaned up documentation, styling issues, JavaScript issues and more. We added the ability for you to use style= in your form elements. We did a lot of internal documentation as well as some external.

Packages may become an internal feature which means we will have to look at the main features of Aisis that make it a revolutionary product for the user as both a developer and a regular user and go from there.

There has been fixes to the admin checks so that notices for the update only show on the update page. We added a new option and a new option tab called Misc. Options  and the new option allows for people to show related posts on there posts, this does not apply to Aisis Articles and Essays. We have removed all the color picker and deleted the whole admin section where we are calling scripts and telling them to load across the whole admin side. We also fixed issues where scripts like css were loading across the admin side.

We added a new email subscription which takes in FeedBurner id or MailChimp and we give you the option to put in a quick blurb. We throw an error if the user registers both.

We use the jtweetanywhere.js which comes with some really nice options that we inocrperated into your twitter widget.

Aisis is coming close to a beta soon :D