This build of Aisis changes the name of the exception file to that of the standard. Class-FileName-Exception.php then we add a new one which is the multisite exception (may be removed). we are also thinking of removing the "override" exception.
This build also introduces MultiSite (Aisis-MultiSite) which essentially allows you to netwrok activate Aisis, then in your appearances section for each site, activate that theme. doing so causes a custom-blog_id folder with associated files to be created. (this is buggy please test).

We have chmod(ed) files and folders in this Aisis-Multisite to that of 0755.

We have also fixed other misc bugs.