Changes were made to the create function for view
Aisis View had a spelling mistake in its create_form_element and thus we
had to go through almost all the files to change this. We also made some
other changes and cleaned up some code. The ability to embed code in
your posts no is done through <pre></pre> instead of
[shortCode][/shortCode] - we still have a handy "insert into post" for
this.