Changes were made, fixes and a new class
Aisis seen a lot of small fixes and changes today. It also saw the
addition of the new Aisis Class: Aisis Custom Loader which is in Aisis
Core package and should be called in the custom function or a custom
package or child theme to call specific loading  of specific files such
as Aisis File handling.