Mass Changes
There have been lots of changes. for one there is now no more "required"
folder in the Modules folder, this has been split up into a lib (for css
and java script) and a images folder and placed in the root of the
adminpanel folder.

All, or most, of the forums now use the cleaner version of Aisis View
and Aisis View has had changes made right down to the interface. now
create_aisis_form_element is doens with just a type of element: input,
text area, select and a array of attributes. type (button, checkbox) is
apart of the array of attributes.

some forums have not moved over to the new refactored way of doing
things and these will be done. Right now documentation needs to be
updated.