Aisis Update Changes
Updates to the update method were made. For Aisis, plus we updated the
interface. We added two new methods that one, check if any files need to
be deleted, and two delete those files. in our case it deletes the Aisis
Core files except those at the root level and one in the template
folder.