Lots of clean up has happened
We are cleaning up a lot of the files in the template section and
through out to make beta 3 more stable, better to use and so on.

We suggest that if you are on a private server to use 777 - else we
suggest you use 755.

By default we use 755 for the custom folder custom css, php and js -
this only applies to network blogs.