We have made lots of changes
We have made a lot of changes to how file hendling class works and to
how the bbpress check is done to see if the plugin is installed or not.
We have added and removed methods and cleaned methods up. now we are to
fix the fall out of this clean up.