We have the options working
So the options are working again, except in WP_DEBUG = true we are
finding that the arrays that are not set prior to or even after clicking
the check, uncheck option are presenting a string offset 0 meaning the
array has nothing in it. that may be able to be solved in a later
commit, currently we have the sidebar, slider and the mini feeds options
up.