Had to do some unconvential fixes
There are a lot of notices telling you such and such is undefined, they
are just that, notices. Other then that the theme template builder
package is coming together - there is no documentation or proper format
at this time. How ever as things become more stable there will be -
continued testing to make sure everythig is working.

Currently we have not tested the  options table with this new addition
how ever we are working on making sure everything is up to speed and
that the options are in the right place. :D

Our options code is simple: $options = $input, this should be looked
into as it seems to simple...