Cleaned up and added new features.
Theres a new cleaner icon set being used for the admin panel, most things about it work properly.

The forms now need to be echoed and there is a new way of getting the label, you have to set up the label attributes in your element array.

We have a better way of rendering view files by just passing in the name and we look for it in the registered paths.