Changed the documentation.
We have fixed the admin issues or at least some of them. The main ones
we have fixed is the tabbing, we have fixed some of the pop up hovers,
we have also changed the footer text tab to misc and allow you to turn
off the twitter boostrap admin theme.

We have also fixed the issue with checkboxes.